Add perform_skipped event type
We currently raise errors when the resque-lock-timeout plugin skips a job to avoid resque-clues logging it as performed.

To avoid that behavior, and make the logs more accurate, a job skip will set `perform_skipped` on the current thread (outside this plugin). And if set, resque-clues will log a `perform_skipped` event.

Removed pry-debugger from the gemspec since it was breaking the bundle, 10 years old, and just a development dependency.